ref: major overhaul of indexing, part 1 of 2-3
This is the first part of a major refactoring overhaul of how we handle
indexing, input-markdown and output-HTML pathing, and sidebar
generation. This is still a work in progress, and most functionality has
been commented out for debugging/testing.

Major changes:

- `update_page_index.py` has been renamed to `create_indices.py`. This
new file now creates two indices:

1. "hier_index", which is the same as the old "index.json". Its content
has not changed. Instead of being saved to file every time and then
loaded later during sidebar creation, it creates and returns the
dictionary instead, and that is passed to sidebar for usage. I'm not
sure if we need to save it as a file at all, since its value can always
be easily debugged live, and we were not doing any kind of checking to
skip its loading before (i.e. if we have to reconstruct it every time,
then there's no gain from saving a copy to disk).

2. "flat_index". This will eventually replace "ordered_links" and is
greatly expanded from what "ordered_links" does. This is treated like
"hier_index" as an explicit "generate_page_html" function, but will replace
much functionality/data of the rest of the code. This index includes the
ordering of legitimate "pages" (i.e. not "sections"), and for every page
contains a dictionary containing:

A. title of the page
B. absolute path of the "input" markdown file
C. absolute path of the "output" HTML file (including processing
dev_build, and including creation of any parent directories to this
not-yet-created file)
D. "website-root"-relative path of the "output" HTML file (including processing
dev_build)

The next major step is using "flat_index" to replace ordered_links and
output path/filename lookups throughout the rest of generate_page_html.

Minor changes:
- Location of the saved index files has been moved to inside "scripts"
from the root
- Many functions and some script files have been renamed