ref: indexing overhaul part 3 of 3
Among other minor changes, including adding a new optional
"save_indices" CLI arg that does what it sounds like, this finishes
applying the usage of the new "flat_index" to the remaining code, which
was mostly the "ordered" links used in the per-page footers.

Everything appears to be working correctly, including that the sidebar
and footer navigation uses "dev" or "content" properly.

Next up is general cleaning and documenting.