ref: attn: remove non-exec hash json recording
This comments out what I think is something that may have been added by
accident. Compare this line to
https://github.com/dylansdaniels/textbook/blob/31dbaf09d6d854a4e8527024fa99f2b415820192/scripts/convert_notebooks.py#L923
.

What was originally going on at the time of the above commit is the
following:
`notebook_was_run` was used as whether or not the notebook was attempted
to be run at all, regardless of if it succeeded (I have changed this
variable to `execution_initiated`). `notebook_executed` was used as
whether or not the notebook successfully finished its execution (I have
changed this variable to `execution_successful`).

In the control flow, which has not changed at all, in this particular
case, we are in the case where `notebook_was_run`/`execution_initiated`
is False, meaning no execution was attempted at all. Also in the
original version, `dev_build` would have been activated, meaning we're
in a dev state. However, even if we were in a dev state, but if the
notebook execution was never initiated/attempted, then in this case, I
do not think it makes sense to record the commit hash of the current HNN
version into the notebook. If we did, it would imply that the last dev
execution was using this commit, but that is not the case.