doc: Add large synopsis of entire building program
This adds a large description of the building program as a whole to the
CLI for 'build.py' such that it is output when calling

python build.py --help

This includes description of the assumed file structure (which has many,
many assumptions), the overall code path, some examples, and the
important options along the way.

Minor: this also adds a docstring for 'get_hnn_commit_hash', renames the
'write_standalone_html' variable to 'save_standalone_nb_html', and makes
a new CLI argument of a similar name. This documentation was first
generated with Claude AI before undergoing *heavy* modification of every
word.