ref: ref all post-val vers checks to is_dev_build
This partially reverts our post-hnn_commit_hash-validation code's checks
of the "code-version" to use a new var like in the original, called
"is_dev_build". The distinction between a "dev" and non dev build is now
fully documented, apparent, and set in 'build.py'. "is_dev_build" is
also a proper boolean flag (true/false only) and the actual commit hash
to use (when necessary) is handled separately.

More refactoring: this also renames the hnn-core commit hash module and
splits its functions, so that the 'get_hnn_commit_hash' function is ONLY
getting the installed version, and the rest of the code is validation of
the installed version against various other versions.