fix: proper nb JSON output loading
This fixes what I believe is a bug introduced in the original
"dev_build" changes.

Here
https://github.com/dylansdaniels/textbook/blob/31dbaf09d6d854a4e8527024fa99f2b415820192/build.py#L114
there is a bug at the later step when the final HTML page is being
built, and when the HTML content of the notebook (from inside its JSON
output file) is being loaded. In this case, because the JSON output file
was being loaded using unique code (different from the JSON output
loading code used at an earlier step inside
`scripts/convert_notebooks.py` see here
https://github.com/dylansdaniels/textbook/blob/31dbaf09d6d854a4e8527024fa99f2b415820192/scripts/convert_notebooks.py#L459
), only the "content" version of the JSON output file was being loaded,
and never the "dev" version. This meant that for "dev" final-page HTML
output containing a notebook's contents, the wrong JSON output was being
used to generate the page.

This has been fixed in that both instances of loading the JSON output
files have been changed to use the same function. Also, the later step
has extra configuration added in order to reconstruct the path of the
desired JSON output file location (which differs in the "dev") case.

More generally, this is another instance of where the complexity
introduced by trying to accomodate the different "dev" vs "stable" file
output locations is a lot, and maybe too much. Perhaps in the future we
can consider changing how he handle a separate "dev" build on the
filesystem on a fundamental level. For example, for a "dev" build, we
could make a full copy in the filesystem of the "content" directory into
the "dev" directory directly, and then replace our "content_path"
variable at the highest level with the "dev" directory.