doc and minor ref: docstrings for remaining funcs
This also moves the output-directory creation step outside of
`create_flat_index` and overtly into the main loop of the HTML page
building code. (If you have to write a ton of comments explaining that a
step is done in a particular location, then maybe it shouldn't be in
that location! I'm dumb and it never should have gone in the index
creation func in the first place).

As is typical, these docstrings were *initially* created with Claude but
then inspected and edited substantially.