ref: change code_version=no_check to CLI arg
This removes the '--code-version' option 'no-check', and instead adds a
standalone CLI argument called '--no-version-validation' that does the
same thing. This allows both "stable" and "dev" builds to EITHER be run
without code-version being checked. (Before, no-check was always a "dev"
build, but this is an improvement since not validating is useful when
inspecting or debugging a local stable build).

This also replaces the confusing argparse not-recommended use of bool in
favor of `action="store_true"`. This is the recommended way to treat CLI
arguments as "presence flags" so to speak, albeit the naming is a little
confusing (the default is False for "store_true").